Extract file-manager helpers into sshpilot/file_manager/ subpackage
First phase of decomposing the 8854-line file_manager_window module.
Move the low-coupling helper layers out:

- file_manager/portal_docs.py — XDG document-portal handling and the
  granted-folders config (DOCS_JSON, _grant_persistent_access,
  _lookup_document_path, _load_first_doc_path, _pretty_path_for_display,
  …).
- file_manager/format_utils.py — _human_size, _human_time,
  _mode_to_str, _mode_to_octal.
- file_manager/remote_walk.py — paramiko helpers _sftp_path_exists,
  stat_isdir, walk_remote.
- file_manager/exceptions.py — TransferCancelledException.

file_manager_window.py re-imports the public names from
sshpilot.file_manager so existing callers (window.py, text_editor.py,
sftp_utils.py, file_manager_integration.py, manual SFTP scripts) and
tests keep working unchanged.

The heavy widget classes (AsyncSFTPManager, FilePane,
FileManagerWindow, SFTPProgressDialog, PropertiesDialog) still live in
file_manager_window.py and will move in follow-up PRs.

Refs #11
