+++ sshpilot/file_manager/exceptions.py
@@ -1,0 +1,8 @@
+"""File-manager-specific exception types."""
+
+from __future__ import annotations
+
+
+class TransferCancelledException(Exception):
+    """Exception raised when a transfer is cancelled."""
+    pass
